# src/main/java/org/onecode/filemanagementapi/dto/response/CourseDTO.java
package org.onecode.filemanagementapi.dto.response;

import lombok.Data;

import java.util.List;

@Data
public class CourseDTO {
    private String id;
    private String name;  // Only include the necessary fields
}